+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 - Price tool: CryptoCompare API rate limit reached returns 200 OK. ([#381](https://github.com/BittyTax/BittyTax/issues/381))
 - Accounting tool: fix AttributeError exception when decoding cell color.
 - Accounting/Price tool: Fix CoinGecko API error "Your request exceeds the allowed time range".
+- Coinbase parser: fix missing warning for advanced trades which are crypto-to-crypto.
+- Coinbase parser: fixed currency conversion when data file is not in local currency.
 ### Added
 - Accounting tool: new PDF report format.
 - Config: added classic_report parameter to switch to legacy PDF report format.